Backend_Main_SpringBoot.  Añadidos modelos de domain y entities JPA para Tecnificaciones y Subtiposde Tecnificación

+++ backend_main_springboot/src/main/resources/db/migration/V3__create_core_feature_tables.sql
@@ -1,5 +1,5 @@
 CREATE TABLE tecnificaciones (
-    id int PRIMARY KEY AUTO_INCREMENT,
+    id bigint PRIMARY KEY AUTO_INCREMENT,
     nombre varchar(255),
     slug varchar(255) UNIQUE,
     descripcion text,
